pyproject.toml
[project]
name = "LMOS-config"
version = "0.1.0"
description = "Configuration management for LMOS."
requires-python = ">=3.8"
dependencies = [
    "pydantic >=2.0",
    "pydantic-settings",
    "ruamel.yaml >=0.18.0",
]

[project.urls]
"Homepage" = "https://github.com/LMOS-IO/lmos-config"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]